refactor(services): PR-3.1 tagging/ 子包搬迁 — 8 文件 + sys.modules 别名 shim
将 services/ 平铺 8 个 tagger 相关文件归到 services/tagging/ 子包：
  tagger.py          → tagging/base.py
  caption_format.py  → tagging/caption_format.py
  caption_snapshot.py → tagging/caption_snapshot.py
  onnx_tagger_base.py → tagging/onnx_base.py
  wd14_tagger.py     → tagging/wd14.py
  cltagger_tagger.py → tagging/cltagger.py
  llm_tagger.py      → tagging/llm.py
  joycaption_tagger.py → tagging/joycaption.py

旧路径全部留 shim 文件，但 **不是普通 re-export**：用
`sys.modules[__name__] = _real` 直接把旧模块对象别名到新模块，让历史
`monkeypatch.setattr(llm_tagger, "_RequestRateLimiter", ...)` 等访问私有 / 隐式
导入子模块的代码与未拆分前行为完全一致。

git mv 保 rename 检测。
回归：route snapshot ✓ / import smoke ✓ / tagger+caption+llm+cltagger 全部测试 149 passed。
